feat: add relations between User and Record models

+++ routes/modules/records.js
@@ -13,8 +13,10 @@ router.get('/new', async (req, res) => {
 // 新增支出
 router.post('/new', async (req, res) => {
   try {
+    const userId = req.user._id
     const { name, date, categoryId, amount } = req.body
     return Record.create({
+      userId,
       name,
       date,
       categoryId,
@@ -30,9 +32,10 @@ router.post('/new', async (req, res) => {
 // edit page
 router.get('/:id/edit', async (req, res) => {
   try {
+    const userId = req.user._id
     const categories = await Category.find().lean()
     const _id = req.params.id
-    const record = await Record.findOne({ _id }).lean()
+    const record = await Record.findOne({ _id, userId }).lean()
     const category = await Category.findOne({ _id: record.categoryId }).lean()
     record.categoryName = category.name
     record.date = record.date.toISOString().slice(0, 10)
@@ -45,10 +48,11 @@ router.get('/:id/edit', async (req, res) => {
 // 修改支出
 router.put('/:id', async (req, res) => {
   try {
+    const userId = req.user._id
     const _id = req.params.id
     const { name, date, categoryId, amount } = req.body
     return Record.findOneAndUpdate(
-      { _id },
+      { _id, userId },
       {
         name,
         date,
@@ -65,7 +69,8 @@ router.put('/:id', async (req, res) => {
 // 刪除功能
 router.delete('/:id', (req, res) => {
   const _id = req.params.id
-  return Record.findOneAndDelete({ _id })
+  const userId = req.user._id
+  return Record.findOneAndDelete({ _id, userId })
     .then(() => res.redirect('/'))
     .catch(err => console.log(err))
 })
